+++ src/test/java/inflearn/springadvance/proxy/advisor/AdvisorTest.java
@@ -3,8 +3,12 @@ package inflearn.springadvance.proxy.advisor;
 import inflearn.springadvance.proxy.common.advice.TimeAdvice;
 import inflearn.springadvance.proxy.common.service.ServiceImpl;
 import inflearn.springadvance.proxy.common.service.ServiceInterface;
+import java.lang.reflect.Method;
 import lombok.extern.slf4j.Slf4j;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.aop.ClassFilter;
+import org.springframework.aop.MethodMatcher;
 import org.springframework.aop.Pointcut;
 import org.springframework.aop.framework.ProxyFactory;
 import org.springframework.aop.support.DefaultPointcutAdvisor;
@@ -25,5 +29,57 @@ public class AdvisorTest {
 
         proxy.find();
         proxy.save();
+    }
+
+    @Test
+    @DisplayName("포인트컷을 직접 만들어보기")
+    void adviserTest2() {
+        ServiceInterface target = new ServiceImpl();
+        ProxyFactory proxyFactory = new ProxyFactory(target);
+
+        DefaultPointcutAdvisor advisor = new DefaultPointcutAdvisor(new MyPointcut(),
+                new TimeAdvice());
+
+        proxyFactory.addAdvisor(advisor);
+        ServiceInterface proxy = (ServiceInterface) proxyFactory.getProxy();
+
+        proxy.find();
+        proxy.save();
+    }
+
+    static class MyPointcut implements Pointcut{
+
+        @Override
+        public ClassFilter getClassFilter() {
+            return ClassFilter.TRUE;
+        }
+
+        @Override
+        public MethodMatcher getMethodMatcher() {
+            return new MyMethodMatcher();
+        }
+    }
+
+    static class MyMethodMatcher implements MethodMatcher{
+
+        public static final String methodName  = "save";
+
+        @Override
+        public boolean matches(Method method, Class<?> targetClass) {
+            boolean result = method.getName().equals(methodName);
+            log.info("포인트컷 호출 method={},  targetClass={}", method.getName());
+            log.info("포인트컷 결과 result={}", result);
+            return result;
+        }
+
+        @Override
+        public boolean isRuntime() {
+            return false;
+        }
+
+        @Override
+        public boolean matches(Method method, Class<?> targetClass, Object... args) {
+            return false;
+        }
     }
 }