+++ CHANGELOG.md
@@ -191,3 +191,6 @@ proxy.call() 하면 이제 프록시 기능이 적용되는 것.
 ---
 
 ## 어드바이저(Adviser)
+- advisortest1() : 어드바이저 (Pointcut.TRUE)
+- advisortest2() : 포인트컷을 직접 만들어 적용
+